Move main to a normal module
